python/pyproject.toml
[tool.poetry]
name = "stochastic-planning"
version = "0.1.0"
description = "Support code for stochastic planning research"
authors = ["CAB <cabemailbox@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.14"
neo4j = "^5.27.0"
pyyed = "^1.5.0"
networkx = "^3.4.2"
beautifulsoup4 = "^4.13.3"
lxml = "^5.3.1"
requests = "^2.32.4"
pyvis = "^0.3.2"
websocket-client = "^1.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
